pyproject.toml
[build-system]
requires = ["setuptools>=57.0.0", "wheel>=0.37.0", "Cython>=0.29.24"]
build-backend = "setuptools.build_meta"

[project]
name = "pensionfi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "cython>=0.29.24",
    "fastapi==0.104.1",
    "fpdf==1.7.2",
    "fpdf2>=2.8.2",
    "motor==3.1.1",
    "numpy>=1.21.0",
    "pydantic>=2.7.4,<3.0.0",
    "pydantic-settings==2.1.0",
    "pymongo==4.3.3",
    "python-dotenv==1.0.0",
    "setuptools>=57.0.0",
    "uvicorn==0.24.0",
    "wheel>=0.37.0",
    "jinja2>=3.1.5",
    "pdfkit>=1.0.0",
    "weasyprint>=64.0",
    "llama-index>=0.11.2",
    "openai>=1.61.1",
    "llama-index-embeddings-voyageai>=0.2.2",
    "langchain>=0.2.5,<0.3.0",
    "langchain-community>=0.2.5",
    "pinecone-client>=5.0.1",
    "voyageai>=0.2.4",
    "pyjwt>=2.10.1",
    "email-validator>=2.1.0",
]

[tool.setuptools]
packages = ["app"]
